Support IPCTNL_MSG_CT_GET_STATS
