+++ src/ctnetlink/nlas/mod.rs
@@ -1,3 +1,5 @@
 // SPDX-License-Identifier: MIT
 
+mod ct_attr;
 pub mod flow;
+pub mod stat;
