+++ src/ctnetlink/nlas/stat/mod.rs
@@ -1,0 +1,3 @@
+// SPDX-License-Identifier: MIT
+
+pub mod nla;
